refactor: move activity, governance, specification boundary+control to domain modules (Phase 1)
Move BCE layers into self-contained domain modules:
- activity-domain: controller, view, and 3 tests
- governance-domain: engine, service, controller, view, and 3 tests
- specification-domain: controller, view, and entity test
Add domainTestDeps for zio-test in domain modules.
Update shared-web views to import ActivityView from activity.boundary.
