+++ modules/activity-domain/src/test/scala/activity/boundary/ActivityViewSpec.scala
@@ -1,4 +1,4 @@
-package shared.web
+package activity.boundary
 
 import java.time.Instant
