+++ modules/shared-web/src/main/scala/shared/web/SdlcDashboardView.scala
@@ -3,6 +3,7 @@ package shared.web
 import java.time.Instant
 import java.time.temporal.ChronoUnit
 
+import activity.boundary.ActivityView
 import activity.entity.ActivityEvent
 import scalatags.Text.all.*
 import sdlc.entity.*
