+++ modules/specification-domain/src/test/scala/specification/entity/SpecificationSpec.scala
